requirements.txt
colour-science==0.4.4
numpy==1.24.4
Pillow==9.2.0
Pygments==2.12.0